fs-tests: test_1: fix compilation warnings
Fix the following compilation warnings:
test_1.c: In function ‘test_1’:
test_1.c:67:4: warning: format ‘%llu’ expects type ‘long long unsigned int’, but argument 3 has type ‘uint64_t’
test_1.c:88:3: warning: format ‘%llu’ expects type ‘long long unsigned int’, but argument 3 has type ‘uint64_t’

The fix is to cast the argument with (unsigned long long).

+++ tests/fs-tests/simple/test_1.c
@@ -64,7 +64,7 @@ void test_1(void)
 	/* Create files to file up the file system */
 	for (block = 1000000, i = 1; ; block /= 10) {
 		while (i != 0) {
-			sprintf(name, "fill_up_%llu", i);
+			sprintf(name, "fill_up_%llu", (unsigned long long)i);
 			actual_size = tests_create_file(name, block);
 			if (actual_size != 0)
 				++i;
@@ -85,7 +85,7 @@ void test_1(void)
 	i -= 1;
 	CHECK(tests_count_files_in_dir(".") == i);
 	for (; i > 0; --i) {
-		sprintf(name, "fill_up_%llu", i);
+		sprintf(name, "fill_up_%llu", (unsigned long long)i);
 		tests_check_filled_file(name);
 		tests_delete_file(name);
 	}
